requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
orjson==3.9.15
packaging==23.2
passlib==1.7.4
peewee==3.17.1
pluggy==1.4.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pytest==8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rabbitmq==0.2.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
